requirements.txt
python-dotenv
sentry-sdk
requests
fastapi
uvicorn
pydantic
